feat(e2e): split bulk-import test into two files
