+++ workspaces/bulk-import/e2e-tests/tests/specs/bulk-import-shared.ts
@@ -1,0 +1,6 @@
+export const WAIT_OBJECTS = {
+  muiLinearProgress: 'div[class*="MuiLinearProgress-root"]',
+  muiCircularProgress: '[class*="MuiCircularProgress-root"]',
+};
+
+export const GITHUB_ORG = "janus-qe";
